Added default react app
Used the basic npm module create-react-app to create a default
react app in react-app folder in the root of the project. Enter the
react-app folder 'cd react-app' then use the 'npm start' command
to see an example of the basic project.

As of now the react app doesn't do anything importnat but could
be exported in the future as part of the project.